refactor: modularize app registration
